Change server to handle client who may close status channel
As well as closing the channel the client might also be in a position
where they cannot renew for long periods of time either.

We now need to keep going when the status channel closes. We
instead introduce a new api /api/:channel/done which the client
throws at the server as its about to exit.

Renews can now happen less frequenly, and we actually cause
the tolken to expire at 4:00am the following day to the point at
it is created.  However as an added protection we need to
stop a long running recording UNLESS it the token is renewed.
We handle that by having a 4 hour timeout on records, except that
the timeout is reset on a renew..