Index.html to load the service worker.
The service worker caches static resources but lets through all
/api calls.

There is one call it treats differently and that is /subscribeid.  The response
to this is cached at service worker installation and will be served from cache
thereafter, essentially giving use a common but unique client id for every
browser client out there.  Thus it should now be possible (except for the
done api command) to restart a client from where it left off.