Fixes #13
It turns out that the order updated really matters, because if after setting a new
state in target, this will trigger a new update BUT it will also be processed
this run through if changes to state are below it and also has a pending
change.  The solution is to put state at the top (which only changes itself)
then target, whih if it changes state will not have that new state processed
until the next run through.

Similarly with the others - although they only effect state and target, so their
order doesn't really matter